No more recursion!
Added REQUEST_PUSH_TO_TASK_SUCCESS because using REQUEST_EDIT_TASK_SUCCESS was kind of a hacky way of doing it in the first place. This action/reducer will push to the files array in the redux store rather than replacing the whole files array whenever you want to add a file. This functionality should also be usable in the future for any task property that is an array.